bug_fix | autotest | surf | post | multi-component systems
wrong counting of the #atoms in both the surface structure and the equi structure, when computing the E per-atom of the corresponding two systems. 
was giving wrong surf-E for multi-component cases, and was giving "nan" if the number of atoms of the first elements is zero.
induced by using the  #atoms of the first elements as the total #atoms.