Update gen.py | compatibility with VASP (#529)
for INCAR param of VASP, NSW = 0 and 1 both lead to the output of 1 converged SCF, thus 1 valid frame of labeled data. 
NSW = 0 (actually single-point calculation) was not supported to be used as the "md_incar".
Though, the previous limitation might be reasonable in semantics, this update just support the practical branching in VASP settings and eliminates the annoying exception for users.