Fix UTs destroyed by pymatgen's update (#1471)
There are 2 changes in pymatgen which destroy the previous unittests:

1. https://github.com/materialsproject/pymatgen/pull/3542
Previously we set the type of DummySpecies to "Type", which is forbidden
by pymatgen through this PR. Here we replace "Type" with "X".

2. https://github.com/materialsproject/pymatgen/pull/3561
The functions from_string and get_string have been replaced with
from_str and get_str. We should keep up to date with these changes.