Bugfix for pimd: sorting model_devi files (#1470)
The previous regular expression did not sort the multiple model_devi.out
files in pimd correctly. As a result, the correspondence between the
model_devi files and the trajectory files was wrong. Since the wrong
configurations were picked, the model_devi accuracy cannot be improved
to a very high level when pimd is used. This PR fixes the bug.

---------

Signed-off-by: Yifan Li李一帆 <yifanl0716@gmail.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Jinzhe Zeng <jinzhe.zeng@rutgers.edu>